move CompilationDatabaseEntry to the proper namespace
